requirements.txt
aiohttp
aiosignal
async-timeout
attrs
Brotli
certifi
cffi
charset-normalizer
discord.py
frozenlist
greenlet
idna
multidict
mutagen
pycparser
pycryptodomex
PyNaCl
requests
SQLAlchemy
typing_extensions
websockets
yarl
youtube-dl
yt-dlp
